Compiles But Crashes

+++ ESP32_Client_V0.1/src/Instruments/InterruptTimer.h
@@ -4,7 +4,9 @@
  */
 #pragma once
 
+#include <stdint.h>
+
 class InterruptTimer {
 public:
-    static void initialize(unsigned long microseconds, void (*isr)());
+    static void initialize(uint32_t microseconds, void (*isr)());
 };
